requirements.txt
httpx>=0.28.1
python-dotenv>=1.2.1
pydantic>=2.0.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
pyodbc>=5.0.0
apscheduler>=3.10.0
asyncio
pytz
